Function call handling.
Relatively conservative argument dependency detection:
Mark all defined "parameter" registers as used.
For now, all potential return registers are marked as defined after call
so they may be incaccurately added to the data flow graph. The reason
this is handled that way is to be certain that DFG equivalence makes
sure that the functions are equivalent (although I this is true no
matter the exact assumptions, as long they are the same between the
functions being compared)